[FEAT] 게시글 수정 및 삭제 기능 완성 (#106)
* chore: 메서드이름수정

* fix: Article도메인 FE작업 복구

* refactor: 게시글 수정 및 삭제 내부로직 개선

* feat: CustomException 핸들링 기능 추가

* refactor: 게시글 삭제 내부로직 개선
-삭제에 실패했을 경우 핸들링 로직 추가

* refactor: 게시글 삭제 내부로직 개선
-삭제에 실패했을 경우 핸들링 로직 추가

* chore: app.js 오타수정

* refactor: 게시글종류 판별 기능 추가
-문자열 입력값으로 게시글 종류 판별 메서드 구현
-클래스 더 명확하게 이름 수정

* chore: 오타수정

* feat: Article, User 도메인 연관관계 매핑 적용

* feat: 게시글 작성 페이지 Article, User 도메인 연관관계 매핑 적용 및 게시글 유형 선택 기능 추가

* style: 게시글 단일조회 페이지 게시글 유형별 뱃지 디자인 수정

* feat: 게시글 단일조회 페이지에서 작성자 표시기능 추가

* style: 게시글 목록조회에서 게시글 종류 표시뱃지 스타일 변경

* refactor: 게시글 삭제 시 작성자만 삭제할 수 있도록 수정

* refactor: 게시글수정 내부로직 개선
-patchDto 추가
-수정 요청시 작성자 아닐경우 에러발생

* style: 게시글작성페이지 버튼디자인 변경

* feat: 게시글 수정 페이지 추가

* feat: 게시글 수정 및 삭제 버튼 추가
-작성자만 볼 수 있는 버튼
-수정버튼 클릭 시 수정페이지로 이동
-삭제버튼 클릭 시 게시글 삭제

* refactor: 게시글 수정시 PUT 메서드 쓰도록 리팩토링

+++ BE/src/main/java/com/myapp/warmwave/domain/article/service/ArticleService.java
@@ -2,6 +2,7 @@ package com.myapp.warmwave.domain.article.service;
 
 import com.myapp.warmwave.common.exception.CustomException;
 import com.myapp.warmwave.common.main.dto.MainArticleDto;
+import com.myapp.warmwave.domain.article.dto.ArticlePatchDto;
 import com.myapp.warmwave.domain.article.dto.ArticlePostDto;
 import com.myapp.warmwave.domain.article.dto.ArticleResponseDto;
 import com.myapp.warmwave.domain.article.entity.Article;
@@ -12,6 +13,7 @@ import com.myapp.warmwave.domain.article.repository.ArticleRepository;
 import com.myapp.warmwave.domain.category.entity.Category;
 import com.myapp.warmwave.domain.category.service.CategoryService;
 import com.myapp.warmwave.domain.image.service.ImageService;
+import com.myapp.warmwave.domain.user.entity.User;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -61,9 +63,15 @@ public class ArticleService {
     }
 
     @Transactional
-    public Article updateArticle(Long articleId, ArticlePostDto dto, List<MultipartFile> imageFiles) throws IOException {
+    public Article updateArticle(String userEmail, ArticlePatchDto dto) throws IOException {
 
-        Article findArticle = articleRepository.findById(articleId).orElseThrow(() -> new CustomException(NOT_FOUND_ARTICLE));
+        Article findArticle = getArticleByArticleId(dto.getArticleId());
+        User user = findArticle.getUser();
+
+        if (!user.getEmail().equals(userEmail)) {
+            throw new CustomException(USER_ROLE_NOT_EXIST);
+        }
+
         imageService.deleteImagesByArticleId(findArticle.getId());
         articleCategoryRepository.deleteByArticleId(findArticle.getId());
 
@@ -76,7 +84,7 @@ public class ArticleService {
 
         //추후 세터를 삭제하는 방향을 생각해보아야함
         findArticle.applyPatch(dto, articleCategoryRepository.findByArticleId(findArticle.getId()));
-        findArticle.setArticleImages(imageService.uploadImages(findArticle, imageFiles));
+        findArticle.setArticleImages(imageService.uploadImages(findArticle, dto.getFiles()));
 
         return articleRepository.save(findArticle);
     }
@@ -99,7 +107,14 @@ public class ArticleService {
     }
 
     @Transactional
-    public void deleteArticle(long articleId) {
+    public void deleteArticle(String userName, long articleId) {
+        Article article = getArticleByArticleId(articleId);
+        User user = article.getUser();
+
+        if (!user.getEmail().equals(userName)) {
+            throw new CustomException(USER_ROLE_NOT_EXIST);
+        }
+
         imageService.deleteImagesByArticleId(articleId);
         articleCategoryRepository.deleteByArticleId(articleId);
         articleRepository.deleteById(articleId);
